+++ tests/test_train_bert.py
@@ -10,6 +10,7 @@ import torch
 from transformers import DistilBertConfig, DistilBertForSequenceClassification
 
 import src.models.bert as model_bert_module
+import src.tokenization.bert as bert_tokenization_module
 from src.models.bert import DistilBERTSentiment
 from tiny_tokenizer import TinyTokenizer
 from src.train_bert import (
@@ -77,10 +78,21 @@ def _patch_hf(monkeypatch):
         classmethod(model_from_pretrained),
     )
     monkeypatch.setattr(
-        train_bert_module.AutoTokenizer,
+        bert_tokenization_module.AutoTokenizer,
         "from_pretrained",
         classmethod(tokenizer_from_pretrained),
     )
+
+
+def test_dataset_bert_wrapper_exports_tokenization_api():
+    import src.dataset_bert as compat
+
+    assert compat.BertReviewDataset is bert_tokenization_module.BertReviewDataset
+    assert compat.encode_texts is bert_tokenization_module.encode_texts
+    assert compat.load_tokenizer is bert_tokenization_module.load_tokenizer
+    assert compat.make_bert_dataloaders is bert_tokenization_module.make_bert_dataloaders
+    assert compat.make_bert_test_loader is bert_tokenization_module.make_bert_test_loader
+    assert compat.resolve_device is bert_tokenization_module.resolve_device
 
 
 def _fixtures(monkeypatch):
